seperated navbar and sidebar
