--- frontend/src/components/AppDrawer.js
@@ -1,250 +1,0 @@
-import * as React from 'react';
-import { styled, useTheme } from '@mui/material/styles';
-import Box from '@mui/material/Box';
-import MuiDrawer from '@mui/material/Drawer';
-import MuiAppBar from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import Button from '@mui/material/Button';
-import List from '@mui/material/List';
-import CssBaseline from '@mui/material/CssBaseline';
-import IconButton from '@mui/material/IconButton';
-import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import ChevronRightIcon from '@mui/icons-material/ChevronRight';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import ForumIcon from '@mui/icons-material/Forum';
-import MessageIcon from '@mui/icons-material/Message';
-import NotificationsIcon from '@mui/icons-material/Notifications';
-import PlaylistAddCheckIcon from '@mui/icons-material/PlaylistAddCheck';
-import SwapHorizIcon from '@mui/icons-material/SwapHoriz';
-import ChecklistIcon from '../assets/ChecklistIcon';
-import WishListIcon from '../assets/WishlistIcon';
-import PreapprovalIcon from '../assets/PreapprovalIcon';
-import LearningAgreementIcon from '../assets/LearningAgreementIcon';
-
-const drawerWidth = 240;
-
-const openedMixin = (theme) => ({
-  width: drawerWidth,
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: 'hidden',
-});
-
-const closedMixin = (theme) => ({
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: 'hidden',
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up('sm')]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-
-const DrawerHeader = styled('div')(({ theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'flex-end',
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-}));
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})(({ theme, open }) => ({
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
-  ({ theme, open }) => ({
-    width: drawerWidth,
-    flexShrink: 0,
-    whiteSpace: 'nowrap',
-    boxSizing: 'border-box',
-    ...(open && {
-      ...openedMixin(theme),
-      '& .MuiDrawer-paper': openedMixin(theme),
-    }),
-    ...(!open && {
-      ...closedMixin(theme),
-      '& .MuiDrawer-paper': closedMixin(theme),
-    }),
-  }),
-);
-
-export default function AppDrawer() {
-  const sideBarItems = [
-    {
-      id: 0,
-      label: "To Do List",
-      icon: (<PlaylistAddCheckIcon/>),
-    },
-    {
-      id: 1,
-      label: "Checklist",
-      icon: (<ChecklistIcon/>),
-    },
-    {
-      id: 2,
-      label: "Wish List",
-      icon: (<WishListIcon/>),
-    },
-    {
-      id: 3,
-      label: "Preapproval",
-      icon: (<PreapprovalIcon/>),
-    },
-    {
-      id: 4,
-      label: "Learning Agreement",
-      icon: (<LearningAgreementIcon/>),
-    },
-    {
-      id: 5,
-      label: "Course Transfer",
-      icon: (<SwapHorizIcon/>),
-    }
-  ]
-  const navBarItems = [
-    {
-      id: 0,
-      button: (<Button size="large" color="inherit" sx={{textTransform: 'none'}}> Application </Button>),
-    },
-    {
-      id: 1,
-      button: (<Button size="large" color="inherit" sx={{textTransform: 'none'}}> Contacts </Button>),
-    },
-    {
-      id: 2,
-      button: (<Button size="large" color="inherit" sx={{textTransform: 'none'}}> Announcements </Button>),
-    },
-    {
-      id: 3,
-      button: (<Button size="large" color="inherit" sx={{textTransform: 'none'}}> Evaluations </Button>), 
-    }
-  ]
-  const theme = useTheme();
-  const [open, setOpen] = React.useState(false);
-
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
-
-  const handleDrawerClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <Box sx={{ display: 'flex' }}>
-      <CssBaseline />
-      <AppBar position="fixed" open={open} sx={{backgroundColor: '#201F2B'}}>
-        <Toolbar>
-          <IconButton
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            size="large"
-            edge="start"
-            color="inherit"
-            sx={{
-              marginRight: 5,
-              ...(open && { display: 'none' }),
-            }}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Box>
-            <List sx={{display: 'flex', flexDirection: 'row'}}>
-              {navBarItems.map((item) => (
-                <ListItem key={item.id} disablePadding sx={{display: 'block' }}>
-                  {item.button}
-                </ListItem>
-              ))}
-            </List>
-          </Box>
-          <Box sx={{flexGrow: 1}} />
-          <Box>
-            <IconButton
-              size="large"
-              color="inherit"
-            >
-              <ForumIcon/>
-            </IconButton>
-            <IconButton
-              size="large"
-              color="inherit"
-            >
-              <MessageIcon/>
-            </IconButton>
-            <IconButton
-              size="large"
-              color="inherit"
-            >
-              <NotificationsIcon/>
-            </IconButton>
-            <IconButton
-              size="large"
-              color="inherit"
-            >
-              <AccountCircleIcon/>
-            </IconButton>
-          </Box>
-        </Toolbar>
-      </AppBar>
-      <Drawer variant="permanent" open={open}>
-        <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === 'rtl' ? <ChevronRightIcon /> : <ChevronLeftIcon />}
-          </IconButton>
-        </DrawerHeader>
-        <List>
-          {sideBarItems.map((item) => (
-            <ListItem key={item.id} disablePadding sx={{ display: 'block' }}>
-              <ListItemButton
-                sx={{
-                  minHeight: 48,
-                  justifyContent: open ? 'initial' : 'center',
-                  px: 2.5,
-                }}
-              >
-                <ListItemIcon
-                  sx={{
-                    minWidth: 0,
-                    mr: open ? 3 : 'auto',
-                    justifyContent: 'center',
-                  }}
-                >
-                  {item.icon}
-                </ListItemIcon>
-                <ListItemText primary={item.label} sx={{ opacity: open ? 1 : 0 }} />
-              </ListItemButton>
-            </ListItem>
-          ))}
-        </List>
-      </Drawer>
-      <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
-        <DrawerHeader />
-      </Box>
-    </Box>
-  );
-}
